requirements.txt
aiogram==3.*
python-dotenv
undetected-chromedriver